ntos: Server-side work for file system drivers
1. Implement IoRegisterFileSystem()
2. Simplify object parsing and opening. In particular, open routine
   invocation is no longer recursive, meaning that if during an open
   an object has its open routine invoked on it, we no longer invoke
   any further open routines on the opened instance. Only parse
   routines will be invoked from that point on.
3. Implement device object parse, insertion, and open routines for
   volume devices.
4. Move members in the file object that belong to the same on-disk
   file to the file control block of the file object.

+++ private/ntos/inc/mm.h
@@ -163,12 +163,12 @@ static inline BOOLEAN MmCapTreeNodeHasChildren(IN PCAP_TREE_NODE Node)
     return !IsListEmpty(&Node->ChildrenList);
 }
 
-static inline ULONG MmCapTreeNodeChildrenCount(IN PCAP_TREE_NODE Node)
+static inline SIZE_T MmCapTreeNodeChildrenCount(IN PCAP_TREE_NODE Node)
 {
     return GetListLength(&Node->ChildrenList);
 }
 
-static inline ULONG MmCapTreeNodeSiblingCount(IN PCAP_TREE_NODE Node)
+static inline SIZE_T MmCapTreeNodeSiblingCount(IN PCAP_TREE_NODE Node)
 {
     return GetListLength(&Node->SiblingLink);
 }
